requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sse-starlette==1.6.1
redis[asyncio]==5.0.1
fastapi-limiter==0.1.6
celery[redis]==5.3.6
bcrypt==4.1.1
setuptools>=68.0.0
